can edit name for template

+++ backend/se_tools/models/discoveryQuestionTemplate/summary.go
@@ -6,6 +6,6 @@ type Summary struct {
 	ID        string    `json:"id"` // This is the public_id
 	Name      string    `json:"name"`
 	Author    string    `json:"author"`
-	Questions int       `json:"questions"`
+	Questions int64     `json:"questions"`
 	UpdateAt  time.Time `json:"updated_at,omitempty"`
 }
